Allow using view locations in cancelled stock moves.
Since changest 29ab3d9bd918 we allow using view locations in draft and staging
moves making it possible to have a warehouse with a view location as main
storage location. But if the user wanted to cancel a shipment with such location
she would have to change the location of all moves before cancelling which is
very annoying.

issue5891
review29531002