Add moves for product quantities by warehouse
We add a new model based on stock move that compute the cumulative quantity for
each incoming or outgoing move.
The product quantities by warehouse is refined to only get row at the date the
quantity actually change.
We also add company and access rules for both models.
On back-end without window function, the cumulative quantity is the initial
quantity of the day.

issue9389
review297881002