Keep the original context when opening Product Quantities by Warehouse
The PYSON context only works when it is evaluated from keyword action.
When it is evaluated as the result of a wizard, there is no more any active*
keys in the context. So we must not change the context as we already have the
right inherited one from the view.

issue9884
review324561002