Improve moves synchronisation on drop shipment
By using an origin drop as synchronisation key, the code is simplified but also
it allows to follow when products are changed.
This allows also to synchronise lots.
The origin drop is filled on the customer moves when the drop shipment is set
to waiting by using the purchase request as link.

issue11700
review411701003